+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 1.6.x
+
+Features:
+
+ - Options page
+ - Configurable warnigs
+
+Bugfixes:
+ 
+ - New installer code using chrome API
+ - Typo fixed (Thanks to Nele)
+ 
 ## 1.5.x
 
 Features:
@@ -12,4 +24,3 @@ Bugfixes:
  - Persistent storage of warnings
  - Handle case where account information is requested when not logged in
  - No "extension updated" notification in case of bugfix
- - Typo fixed (Thanks to Nele)